// src/dishhive-web/src/app/api/generated/dishhive-api.client.ts
export enum FirstDayOfWeek {
    Monday = "Monday",
    Sunday = "Sunday",
}
// ...
export enum MeasurementSystem {
    Metric = "Metric",
    Imperial = "Imperial",
}